refactor(similarity): split prep orchestration and store
