+++ src/egui_app/controller/similarity_prep/state.rs
@@ -119,4 +119,16 @@ mod tests {
         assert!(!state.skip_backfill);
         assert!(transition.should_enqueue_embeddings);
     }
+
+    #[test]
+    fn finalize_only_starts_from_embeddings_stage() {
+        let mut state = build_initial_state(build_init(false, false, false));
+        assert!(start_finalize_if_ready(&mut state).is_none());
+        assert_eq!(state.stage, SimilarityPrepStage::AwaitScan);
+
+        let mut state = build_initial_state(build_init(true, false, false));
+        let request = start_finalize_if_ready(&mut state).expect(\"finalize request\");
+        assert_eq!(state.stage, SimilarityPrepStage::Finalizing);
+        assert_eq!(request.umap_version, \"v1\");
+    }
 }
